Fixup pointer handling

+++ examples/dump.rs
@@ -11,7 +11,7 @@ fn print_property(parser: &mut Parser, property: &Property) {
         TdhInType::InTypeAnsiString => println!("{:?}", TryParse::<String>::try_parse(parser, &property.name)),
         TdhInType::InTypeUInt32 => println!("{:?}", TryParse::<u32>::try_parse(parser, &property.name)),
         TdhInType::InTypeUInt8 => println!("{:?}", TryParse::<u8>::try_parse(parser, &property.name)),
-        TdhInType::InTypePointer => println!("{:?}", TryParse::<usize>::try_parse(parser, &property.name)),
+        TdhInType::InTypePointer => println!("{:?}", TryParse::<u64>::try_parse(parser, &property.name)),
         TdhInType::InTypeInt64 => println!("{:?}", TryParse::<i64>::try_parse(parser, &property.name)),
         TdhInType::InTypeGuid => println!("{:?}", TryParse::<Guid>::try_parse(parser, &property.name)),
         _ => println!("Unknown {:?}", property.in_type())
